Revert "Anthropic refactoring"
