+++ examples/anthropic/csv_scraper_graph_multi_haiku.py
@@ -1,25 +1,23 @@
-""" 
-Basic example of scraping pipeline using SmartScraper from text
+"""
+Basic example of scraping pipeline using CSVScraperMultiGraph from CSV documents
 """
 
 import os
 from dotenv import load_dotenv
-from scrapegraphai.graphs import SmartScraperGraph
-from scrapegraphai.utils import prettify_exec_info
+import pandas as pd
+from scrapegraphai.graphs import CSVScraperMultiGraph
+from scrapegraphai.utils import convert_to_csv, convert_to_json, prettify_exec_info
 
 load_dotenv()
-
 # ************************************************
-# Read the text file
+# Read the CSV file
 # ************************************************
 
-FILE_NAME = "inputs/plain_html_example.txt"
+FILE_NAME = "inputs/username.csv"
 curr_dir = os.path.dirname(os.path.realpath(__file__))
 file_path = os.path.join(curr_dir, FILE_NAME)
 
-# It could be also a http request using the request model
-with open(file_path, 'r', encoding="utf-8") as file:
-    text = file.read()
+text = pd.read_csv(file_path)
 
 # ************************************************
 # Define the configuration for the graph
@@ -28,27 +26,30 @@ with open(file_path, 'r', encoding="utf-8") as file:
 graph_config = {
     "llm": {
         "api_key": os.getenv("ANTHROPIC_API_KEY"),
-        "model": "anthropic/claude-3-haiku-20240307",
-        "max_tokens": 4000
-        },
+        "model": "claude-3-haiku-20240307",
+        "max_tokens": 4000},
 }
 
 # ************************************************
-# Create the SmartScraperGraph instance and run it
+# Create the CSVScraperMultiGraph instance and run it
 # ************************************************
 
-smart_scraper_graph = SmartScraperGraph(
-    prompt="List me all the projects with their description.",
-    source=text,
+csv_scraper_graph = CSVScraperMultiGraph(
+    prompt="List me all the last names",
+    source=[str(text), str(text)],
     config=graph_config
 )
 
-result = smart_scraper_graph.run()
+result = csv_scraper_graph.run()
 print(result)
 
 # ************************************************
 # Get graph execution info
 # ************************************************
 
-graph_exec_info = smart_scraper_graph.get_execution_info()
+graph_exec_info = csv_scraper_graph.get_execution_info()
 print(prettify_exec_info(graph_exec_info))
+
+# Save to json or csv
+convert_to_csv(result, "result")
+convert_to_json(result, "result")
