+++ examples/anthropic/json_scraper_haiku.py
@@ -1,62 +1,57 @@
 """
-Basic example of scraping pipeline using CSVScraperGraph from CSV documents
+Basic example of scraping pipeline using JSONScraperGraph from JSON documents
 """
 
 import os
 from dotenv import load_dotenv
-import pandas as pd
-from scrapegraphai.graphs import CSVScraperGraph
+from scrapegraphai.graphs import JSONScraperGraph
 from scrapegraphai.utils import convert_to_csv, convert_to_json, prettify_exec_info
-
 load_dotenv()
 
 # ************************************************
-# Read the CSV file
+# Read the JSON file
 # ************************************************
 
-FILE_NAME = "inputs/username.csv"
+FILE_NAME = "inputs/example.json"
 curr_dir = os.path.dirname(os.path.realpath(__file__))
 file_path = os.path.join(curr_dir, FILE_NAME)
 
-text = pd.read_csv(file_path)
+with open(file_path, 'r', encoding="utf-8") as file:
+    text = file.read()
 
 # ************************************************
 # Define the configuration for the graph
 # ************************************************
 
-# required environment variables in .env
-# HUGGINGFACEHUB_API_TOKEN
-# ANTHROPIC_API_KEY
-load_dotenv()
-
 graph_config = {
     "llm": {
         "api_key": os.getenv("ANTHROPIC_API_KEY"),
-        "model": "anthropic/claude-3-haiku-20240307",
+        "model": "claude-3-haiku-20240307",
         "max_tokens": 4000
         },
 }
 
 # ************************************************
-# Create the CSVScraperGraph instance and run it
+# Create the JSONScraperGraph instance and run it
 # ************************************************
 
-csv_scraper_graph = CSVScraperGraph(
-    prompt="List me all the last names",
-    source=str(text),  # Pass the content of the file, not the file object
+json_scraper_graph = JSONScraperGraph(
+    prompt="List me all the authors, title and genres of the books",
+    source=text,  # Pass the content of the file, not the file object
     config=graph_config
 )
 
-result = csv_scraper_graph.run()
+result = json_scraper_graph.run()
 print(result)
 
 # ************************************************
 # Get graph execution info
 # ************************************************
 
-graph_exec_info = csv_scraper_graph.get_execution_info()
+graph_exec_info = json_scraper_graph.get_execution_info()
 print(prettify_exec_info(graph_exec_info))
 
 # Save to json or csv
 convert_to_csv(result, "result")
 convert_to_json(result, "result")
+
