+++ examples/anthropic/scrape_plain_text_haiku.py
@@ -1,21 +1,23 @@
-"""
-Basic example of scraping pipeline using XMLScraperGraph from XML documents
+""" 
+Basic example of scraping pipeline using SmartScraper from text
 """
 
 import os
 from dotenv import load_dotenv
-from scrapegraphai.graphs import XMLScraperGraph
-from scrapegraphai.utils import convert_to_csv, convert_to_json, prettify_exec_info
+from scrapegraphai.graphs import SmartScraperGraph
+from scrapegraphai.utils import prettify_exec_info
+
 load_dotenv()
 
 # ************************************************
-# Read the XML file
+# Read the text file
 # ************************************************
 
-FILE_NAME = "inputs/books.xml"
+FILE_NAME = "inputs/plain_html_example.txt"
 curr_dir = os.path.dirname(os.path.realpath(__file__))
 file_path = os.path.join(curr_dir, FILE_NAME)
 
+# It could be also a http request using the request model
 with open(file_path, 'r', encoding="utf-8") as file:
     text = file.read()
 
@@ -26,31 +28,27 @@ with open(file_path, 'r', encoding="utf-8") as file:
 graph_config = {
     "llm": {
         "api_key": os.getenv("ANTHROPIC_API_KEY"),
-        "model": "anthropic/claude-3-haiku-20240307",
+        "model": "claude-3-haiku-20240307",
         "max_tokens": 4000
         },
 }
 
 # ************************************************
-# Create the XMLScraperGraph instance and run it
+# Create the SmartScraperGraph instance and run it
 # ************************************************
 
-xml_scraper_graph = XMLScraperGraph(
-    prompt="List me all the authors, title and genres of the books",
-    source=text,  # Pass the content of the file, not the file object
+smart_scraper_graph = SmartScraperGraph(
+    prompt="List me all the projects with their description.",
+    source=text,
     config=graph_config
 )
 
-result = xml_scraper_graph.run()
+result = smart_scraper_graph.run()
 print(result)
 
 # ************************************************
 # Get graph execution info
 # ************************************************
 
-graph_exec_info = xml_scraper_graph.get_execution_info()
+graph_exec_info = smart_scraper_graph.get_execution_info()
 print(prettify_exec_info(graph_exec_info))
-
-# Save to json or csv
-convert_to_csv(result, "result")
-convert_to_json(result, "result")
