+++ examples/anthropic/script_multi_generator_haiku.py
@@ -16,7 +16,7 @@ load_dotenv()
 graph_config = {
     "llm": {
         "api_key": os.getenv("ANTHROPIC_API_KEY"),
-        "model": "anthropic/claude-3-haiku-20240307",
+        "model": "claude-3-haiku-20240307",
         "max_tokens": 4000
         },
         "library": "beautifulsoup"
