+++ examples/anthropic/search_graph_haiku.py
@@ -3,31 +3,19 @@ Example of Search Graph
 """
 
 import os
-from typing import List
 from dotenv import load_dotenv
-from pydantic import BaseModel, Field
 from scrapegraphai.graphs import SearchGraph
-
+from scrapegraphai.utils import convert_to_csv, convert_to_json, prettify_exec_info
 load_dotenv()
-
-# ************************************************
-# Define the output schema for the graph
-# ************************************************
-
-class Dish(BaseModel):
-    name: str = Field(description="The name of the dish")
-    description: str = Field(description="The description of the dish")
-
-class Dishes(BaseModel):
-    dishes: List[Dish]
 
 # ************************************************
 # Define the configuration for the graph
 # ************************************************
+
 graph_config = {
     "llm": {
         "api_key": os.getenv("ANTHROPIC_API_KEY"),
-        "model": "anthropic/claude-3-haiku-20240307",
+        "model": "claude-3-haiku-20240307",
         "max_tokens": 4000
         },
 }
@@ -38,9 +26,19 @@ graph_config = {
 
 search_graph = SearchGraph(
     prompt="List me Chioggia's famous dishes",
-    config=graph_config,
-    schema=Dishes
+    config=graph_config
 )
 
 result = search_graph.run()
 print(result)
+
+# ************************************************
+# Get graph execution info
+# ************************************************
+
+graph_exec_info = search_graph.get_execution_info()
+print(prettify_exec_info(graph_exec_info))
+
+# Save to json and csv
+convert_to_csv(result, "result")
+convert_to_json(result, "result")
