+++ examples/anthropic/smart_scraper_haiku.py
@@ -19,7 +19,7 @@ load_dotenv()
 graph_config = {
     "llm": {
         "api_key": os.getenv("ANTHROPIC_API_KEY"),
-        "model": "anthropic/claude-3-haiku-20240307",
+        "model": "claude-3-haiku-20240307",
         "max_tokens": 4000
         },
 }
