+++ examples/anthropic/smart_scraper_schema_haiku.py
@@ -33,7 +33,7 @@ class Projects(BaseModel):
 graph_config = {
     "llm": {
         "api_key": os.getenv("ANTHROPIC_API_KEY"),
-        "model": "anthropic/claude-3-haiku-20240307",
+        "model": "claude-3-haiku-20240307",
         "max_tokens": 4000},
 }
