+++ examples/anthropic/xml_scraper_graph_multi_haiku.py
@@ -26,7 +26,7 @@ with open(file_path, 'r', encoding="utf-8") as file:
 graph_config = {
     "llm": {
         "api_key": os.getenv("ANTHROPIC_API_KEY"),
-        "model": "anthropic/claude-3-haiku-20240307",
+        "model": "claude-3-haiku-20240307",
         "max_tokens": 4000},
 }
